feat(backend): move pagination models to own file
