+++ invokeai/app/services/board_record_storage.py
@@ -5,9 +5,9 @@ from typing import Optional, Union, cast
 
 from pydantic import BaseModel, Extra, Field
 
-from invokeai.app.services.image_record_storage import OffsetPaginatedResults
 from invokeai.app.services.models.board_record import BoardRecord, deserialize_board_record
 from invokeai.app.services.shared.db import SqliteDatabase
+from invokeai.app.services.shared.pagination import OffsetPaginatedResults
 from invokeai.app.util.misc import uuid_string
