+++ invokeai/app/services/session_queue/session_queue_base.py
@@ -18,7 +18,7 @@ from invokeai.app.services.session_queue.session_queue_common import (
     SessionQueueItemDTO,
     SessionQueueStatus,
 )
-from invokeai.app.services.shared.models import CursorPaginatedResults
+from invokeai.app.services.shared.pagination import CursorPaginatedResults
 
 
 class SessionQueueBase(ABC):
